feat: modify super-like btn action
